+++ libft/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: orudek <orudek@student.42madrid.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/11 18:19:42 by orudek            #+#    #+#             */
-/*   Updated: 2023/03/13 14:29:27 by orudek           ###   ########.fr       */
+/*   Updated: 2023/03/13 16:14:11 by orudek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (!n)
